pyproject.toml
[project]
name = "crewai-human-input-interactive"
version = "0.1.0"
requires-python = ">=3.12.9"
dependencies = [
    "chainlit>=2.5.5",
    "click>=8.2.1",
    "crewai",
    "crewai-tools",
    "embedchain==0.1.126",
    "ipykernel==6.29.5",
    "ipython==8.32.0",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "langchain==0.3.18",
    "langchain-community>=0.3.1,<0.4.0",
    "langchain-core>=0.3.34",
    "langchain-openai==0.2.14",
    "loguru>=0.7.3",
    "openai==1.62.0",
    "opentelemetry-api==1.29.0",
    "opentelemetry-exporter-otlp==1.29.0",
    "opentelemetry-sdk==1.29.0",
    "pydantic>=2.10.6",
    "python-dotenv==1.0.1",
    "requests>=2.32.3",
    "tiktoken==0.7.0",
]

[lint]
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293', 'E501']
exclude = ["app/**", "tools/**", ".venv"]
